+++ lib/src/widgets/context_menu_item_widget.dart
@@ -126,7 +126,8 @@ class _ContextMenuItemWidgetState extends State<ContextMenuItemWidget> {
                 if (widget.item.isSubmenu)
                   Opacity(
                     opacity: widget.item.enabled ? 0.6 : 0.3,
-                    child: const Icon(Icons.chevron_right, size: 16),
+                    child: widget.config.submenu.icon ??
+                        const Icon(Icons.chevron_right, size: 16),
                   ),
               ],
             ),